update java docs for existing code
