--- app/src/main/java/com/group80/uoftinder/firebase/StorageDbDownloadable.java
@@ -1,9 +1,0 @@
-package com.group80.uoftinder.firebase;
-
-import androidx.annotation.NonNull;
-
-public interface StorageDbDownloadable<T> {
-    void onStorageDownloadSuccess(T data);
-
-    void onStorageDownloadFailure(@NonNull Exception exception);
-}
