+++ app/src/main/java/com/group80/uoftinder/firebase/storage/ucImageDownloader.java
@@ -1,4 +1,4 @@
-package com.group80.uoftinder.firebase;
+package com.group80.uoftinder.firebase.storage;
 
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
@@ -6,7 +6,16 @@ import android.graphics.BitmapFactory;
 import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageReference;
 
+/**
+ * A downloader to download bitmap images from the firebase cloud storage
+ */
 public class ucImageDownloader extends ucDownloader<Bitmap, byte[]> {
+    /**
+     * Downloads an byte array from the firebase storage, then convert it to an bitmap image
+     *
+     * @param downloadUri the url to download the image
+     * @return the bitmap image
+     */
     @Override
     public Bitmap download(String downloadUri) {
         StorageReference imageReference = FirebaseStorage.getInstance().getReferenceFromUrl(downloadUri);
